pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "exponent-ml"
version = "0.1.2-beta"
description = "Claude Code for ML developers. Build and deploy ML models from your terminal. Accelerate your ML pipeline with Exponent."
authors = [
    {name = "Eden Etuk", email = "edens.etuk@gmail.com"}
]
maintainers = [
    {name = "Eden Etuk", email = "edens.etuk@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
keywords = ["machine-learning", "cli", "ai", "ml", "automation", "modal", "anthropic"]
dependencies = [
    "typer>=0.9.0",
    "anthropic>=0.7.0",
    "boto3>=1.26.0",
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "modal>=0.55.0",
    "PyGithub>=1.59.0",
    "python-dotenv>=1.0.0",
    "inquirer>=3.1.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "clerk-sdk-python>=0.1.0",
    "requests>=2.28.0",
    "rich>=13.0.0",
    "openai>=1.0.0",
    "flask>=2.3.0",
    "flask-cors>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
docs = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "myst-parser>=1.0.0",
]

[project.scripts]
exponent = "exponent.main:app"

[project.urls]
Homepage = "https://github.com/3d3n-ops/exponent-0.1"
Repository = "https://github.com/3d3n-ops/exponent-0.1"
Documentation = "https://exponent-ml.vercel.app/docs"
Issues = "https://github.com/3d3n-ops/exponent-0.1/issues"


[tool.setuptools.packages.find]
where = ["."]
include = ["exponent*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"
